dns_random: Use uint32_t and not unsigned int
Also, make sure the TSIG key generator will always fill up all the bits
in the key.

+++ pdns/dns_random.cc
@@ -210,7 +210,7 @@ void dns_random_init(const string& data __attribute__((unused)), bool force) {
 }
 
 /* Parts of this code come from arc4random_uniform */
-unsigned int dns_random(unsigned int upper_bound) {
+uint32_t dns_random(uint32_t upper_bound) {
   if (chosen_rng == RNG_UNINITIALIZED)
     dns_random_setup();
 
@@ -241,7 +241,7 @@ unsigned int dns_random(unsigned int upper_bound) {
     throw std::runtime_error("Unreachable at " __FILE__ ":" + boost::lexical_cast<std::string>(__LINE__)); // cannot be reached
   case RNG_SODIUM:
 #if defined(HAVE_RANDOMBYTES_STIR) && !defined(USE_URANDOM_ONLY)
-    return static_cast<unsigned int>(randombytes_uniform(static_cast<uint32_t>(upper_bound)));
+    return randombytes_uniform(upper_bound);
 #else
     throw std::runtime_error("Unreachable at " __FILE__ ":" + boost::lexical_cast<std::string>(__LINE__)); // cannot be reached
 #endif /* RND_SODIUM */
@@ -271,7 +271,7 @@ unsigned int dns_random(unsigned int upper_bound) {
       }
   case RNG_ARC4RANDOM:
 #if defined(HAVE_ARC4RANDOM) && !defined(USE_URANDOM_ONLY)
-    return static_cast<unsigned int>(arc4random_uniform(static_cast<uint32_t>(upper_bound)));
+    return arc4random_uniform(upper_bound);
 #else
     throw std::runtime_error("Unreachable at " __FILE__ ":" + boost::lexical_cast<std::string>(__LINE__)); // cannot be reached
 #endif
